moved components to uploads folder
