+++ client/src/pages/dashboard/sections/ViewButtons.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import { Button } from 'antd';
+
 import { View } from 'helpers/constants/enums';
 import {
   vulnerabilityColumnList,
