+++ client/src/pages/dashboard/sections/paginatedTable/ComponentViewRow.tsx
