+++ client/src/pages/dashboard/sections/paginatedTable/VulnerabilityViewRow.tsx
