+++ client/src/pages/dashboard/sections/paginatedTable/index.ts
@@ -1,0 +1,3 @@
+import PaginatedTable from './PaginatedTable';
+
+export default PaginatedTable;
