+++ client/src/pages/dashboard/sections/quickStats/index.ts
@@ -1,0 +1,3 @@
+import QuickStats from './QuickStats';
+
+export default QuickStats;
