+++ client/src/pages/upload/components/CarouselController.tsx
@@ -1,5 +1,4 @@
 import styled from 'styled-components';
-
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { icon } from '@fortawesome/fontawesome-svg-core/import.macro';
