+++ client/src/pages/upload/components/FileUpload.tsx
@@ -1,6 +1,7 @@
 import { useState, useEffect } from 'react';
 import styled from 'styled-components';
 import axios from 'axios';
+
 import { SbomProcessingState } from 'helpers/constants/enums';
 
 interface FileUploadProps {
