+++ client/src/pages/upload/components/SbomProcessor.tsx
