+++ client/src/pages/upload/components/UploadStep.tsx
