+++ client/src/pages/upload/components/index.ts
@@ -1,0 +1,5 @@
+import CarouselBody from './CarouselBody';
+import CarouselController from './CarouselController';
+import SbomProcessor from './SbomProcessor';
+
+export { CarouselBody, CarouselController, SbomProcessor };
